+++ HW/chicolitta/temperature.h
@@ -3,6 +3,6 @@
 
 int spi_init(char filename[40]);
 char * spi_read_lm74(int file);
-double temp_read(void);
+float temp_read(void);
 
 #endif
